move ShardingSphereSchemaDataAlteredEvent to mode.
